requirements.txt
python-telegram-bot==20.3
openai>=1.0.0
python-dotenv
requests
